Removed DefaultFlushEventListener from public API
- Removed the DefaultFlushEventListener from public API - now it resides  in dataservice-write/src folder.
- Updated the corresponding tests which were using this class.

Relates to: OLPEDGE-825
